requirements.txt
twitter
certifi==2017.11.5
chardet==3.0.4
Django==1.11.6
django-multiselectfield==0.1.5
djangorestframework==3.5.4
docutils==0.13.1
geojson==1.3.3
geopy==1.11.0
idna==2.6
libxml2dom==0.5
lxml==3.6.4
lxml-asserts==0.1.1
phonenumbers==8.8.8
Pillow==5.0.0
psycopg2==2.7.3.2
Pygments==2.2.0
PyJWT==1.5.3
pyproj==1.9.5.1
PySocks==1.6.8
python-dateutil==2.5.3
pytz==2017.3
requests==2.18.4
Shapely==1.5.17
simplejson==3.10.0
six==1.10.0
twilio==6.10.0
urllib3==1.22
python-twitter==3.3
oauthlib>=0.6.2
